+++ src/Classes/IP/Address.java
@@ -4,10 +4,10 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class Address {
-    private List<Byte> value;
+    private List<Integer> value;
     private String dotValue;
 
-    public Address(List<Byte> info) {
+    public Address(List<Integer> info) {
         this.value = info;
         this.dotValue = info.stream().map(String::valueOf).collect(Collectors.joining("."));
     }
